new unified route to get batches output /batches/:id/output
